+++ configs/tsup/index.js
@@ -3,6 +3,7 @@ module.exports = {
   format: "cjs",
   clean: true,
   dts: true,
+  bundle: false,
   entry: ["src", "!src/**/*.spec.*"],
   outDir: "build",
 };