+++ packages/common/src/index.ts
@@ -5,4 +5,3 @@ export { default as Response } from "./response";
 export { default as Cache } from "./cache";
 export * as hoprd from "./hoprd";
 export * as utils from "./utils";
-export * as fixtures from "./fixtures";