+++ packages/crypto-bridge/src/web/index.ts
@@ -1,1 +1,2 @@
 export * from "@rpch/crypto/web";
+export { default as init } from "@rpch/crypto/web";