+++ apps/funding-platform/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @rpch/funding-platform
 
+## 0.0.4
+
+### Patch Changes
+
+- @rpch/common@0.1.3
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @rpch/common@0.1.2
+
 ## 0.0.2
 
 ### Patch Changes