+++ packages/common/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @rpch/common
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @rpch/crypto-bridge@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Introduce web compatibility and various improvements.
+- Updated dependencies
+  - @rpch/crypto-bridge@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes