+++ packages/crypto-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rpch/crypto-bridge
 
+## 0.1.3
+
+### Patch Changes
+
+- Patch publishing bug
+
+## 0.1.2
+
+### Patch Changes
+
+- Introduce web compatibility and various improvements.
+
 ## 0.1.1
 
 ### Patch Changes